requirements.txt
aiohttp==3.8.1
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.10
discord.py==1.7.3
idna==3.3
multidict==5.2.0
numpy==1.22.1
pandas==1.3.5
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
typing_extensions==4.0.1
urllib3==1.26.8
yarl==1.7.2
